+++ packages/core/src/models/module-extract.ts
@@ -1,0 +1,13 @@
+/**
+ * Extaction from current module.
+ */
+export class ModuleExtract {
+  /**
+   * Module name.
+   */
+  name: string = '';
+  /**
+   * Module prefix.
+   */
+  path: string = '';
+}